third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/set-outer-html-for-xhtml-expected.txt

Tests DOMAgent.setOuterHTML protocol method against an XHTML document.


Running: testSetUp

Running: testChangeCharacterData
Replacing 'Getting involved' with 'Getting not involved'

Wrapper identity: identity
Event Symbol(AttrModified): div
Event Symbol(AttrModified): div
Event Symbol(AttrModified): div
Event Symbol(AttrRemoved): div
Event Symbol(AttrRemoved): div
Event Symbol(CharacterDataModified): #text
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting not involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event Symbol(CharacterDataModified): #text
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testChangeAttributes
Replacing '<a href' with '<a foo="bar" href'

Wrapper identity: identity
Event Symbol(AttrModified): a
Event Symbol(AttrModified): a
Event Symbol(AttrRemoved): a
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a foo="bar" href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event Symbol(AttrModified): a
Event Symbol(AttrRemoved): a
Event Symbol(AttrRemoved): a
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testRemoveLastChild
Replacing 'Getting involved' with ''

Wrapper identity: identity
Event Symbol(NodeRemoved): #text
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2></h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event Symbol(NodeInserted): #text
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testSplitNode
Replacing 'Getting involved' with 'Getting</h2><h2>involved'

Wrapper identity: identity
Event Symbol(NodeInserted): h2
Event Symbol(NodeInserted): h2
Event Symbol(NodeRemoved): h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting</h2><h2>involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event Symbol(NodeInserted): h2
Event Symbol(NodeRemoved): h2
Event Symbol(NodeRemoved): h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testChangeNodeName
Replacing '<h2>Getting involved</h2>' with '<h3>Getting involved</h3>'

Wrapper identity: identity
Event Symbol(NodeInserted): h3
Event Symbol(NodeRemoved): h2
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h3>Getting involved</h3>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event Symbol(NodeInserted): h2
Event Symbol(NodeRemoved): h3
==========8<==========
<div xmlns="http://www.w3.org/1999/xhtml" id="container" style="display:none">
<p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc.</p>
<h2>Getting involved</h2>
<p id="identity">There are many ways to get involved. You can:</p>
</div>
==========>8==========





Running: testInvalidDocumentDoesNotCrash
PASS: No crash
